+++ oci/identity_tag_default_resource.go
@@ -1,0 +1,233 @@
+// Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
+
+package provider
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform/helper/schema"
+
+	oci_identity "github.com/oracle/oci-go-sdk/identity"
+)
+
+func IdentityTagDefaultResource() *schema.Resource {
+	return &schema.Resource{
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
+		Timeouts: DefaultTimeout,
+		Create:   createIdentityTagDefault,
+		Read:     readIdentityTagDefault,
+		Update:   updateIdentityTagDefault,
+		Delete:   deleteIdentityTagDefault,
+		Schema: map[string]*schema.Schema{
+			// Required
+			"compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+			"tag_definition_id": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+			"value": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+
+			// Optional
+
+			// Computed
+			"state": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"tag_definition_name": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"tag_namespace_id": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"time_created": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+		},
+	}
+}
+
+func createIdentityTagDefault(d *schema.ResourceData, m interface{}) error {
+	sync := &IdentityTagDefaultResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).identityClient
+
+	return CreateResource(d, sync)
+}
+
+func readIdentityTagDefault(d *schema.ResourceData, m interface{}) error {
+	sync := &IdentityTagDefaultResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).identityClient
+
+	return ReadResource(sync)
+}
+
+func updateIdentityTagDefault(d *schema.ResourceData, m interface{}) error {
+	sync := &IdentityTagDefaultResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).identityClient
+
+	return UpdateResource(d, sync)
+}
+
+func deleteIdentityTagDefault(d *schema.ResourceData, m interface{}) error {
+	sync := &IdentityTagDefaultResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).identityClient
+	sync.DisableNotFoundRetries = true
+
+	return DeleteResource(d, sync)
+}
+
+type IdentityTagDefaultResourceCrud struct {
+	BaseCrud
+	Client                 *oci_identity.IdentityClient
+	Res                    *oci_identity.TagDefault
+	DisableNotFoundRetries bool
+}
+
+func (s *IdentityTagDefaultResourceCrud) ID() string {
+	return *s.Res.Id
+}
+
+func (s *IdentityTagDefaultResourceCrud) CreatedPending() []string {
+	return []string{}
+}
+
+func (s *IdentityTagDefaultResourceCrud) CreatedTarget() []string {
+	return []string{
+		string(oci_identity.TagDefaultLifecycleStateActive),
+	}
+}
+
+func (s *IdentityTagDefaultResourceCrud) DeletedPending() []string {
+	return []string{}
+}
+
+func (s *IdentityTagDefaultResourceCrud) DeletedTarget() []string {
+	return []string{}
+}
+
+func (s *IdentityTagDefaultResourceCrud) Create() error {
+	request := oci_identity.CreateTagDefaultRequest{}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if tagDefinitionId, ok := s.D.GetOkExists("tag_definition_id"); ok {
+		tmp := tagDefinitionId.(string)
+		request.TagDefinitionId = &tmp
+	}
+
+	if value, ok := s.D.GetOkExists("value"); ok {
+		tmp := value.(string)
+		request.Value = &tmp
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "identity")
+
+	response, err := s.Client.CreateTagDefault(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response.TagDefault
+	return nil
+}
+
+func (s *IdentityTagDefaultResourceCrud) Get() error {
+	request := oci_identity.GetTagDefaultRequest{}
+
+	tmp := s.D.Id()
+	request.TagDefaultId = &tmp
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "identity")
+
+	response, err := s.Client.GetTagDefault(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response.TagDefault
+	return nil
+}
+
+func (s *IdentityTagDefaultResourceCrud) Update() error {
+	request := oci_identity.UpdateTagDefaultRequest{}
+
+	tmp := s.D.Id()
+	request.TagDefaultId = &tmp
+
+	if value, ok := s.D.GetOkExists("value"); ok {
+		tmp := value.(string)
+		request.Value = &tmp
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "identity")
+
+	response, err := s.Client.UpdateTagDefault(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response.TagDefault
+	return nil
+}
+
+func (s *IdentityTagDefaultResourceCrud) Delete() error {
+	request := oci_identity.DeleteTagDefaultRequest{}
+
+	tmp := s.D.Id()
+	request.TagDefaultId = &tmp
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "identity")
+
+	_, err := s.Client.DeleteTagDefault(context.Background(), request)
+	return err
+}
+
+func (s *IdentityTagDefaultResourceCrud) SetData() error {
+	if s.Res.CompartmentId != nil {
+		s.D.Set("compartment_id", *s.Res.CompartmentId)
+	}
+
+	s.D.Set("state", s.Res.LifecycleState)
+
+	if s.Res.TagDefinitionId != nil {
+		s.D.Set("tag_definition_id", *s.Res.TagDefinitionId)
+	}
+
+	if s.Res.TagDefinitionName != nil {
+		s.D.Set("tag_definition_name", *s.Res.TagDefinitionName)
+	}
+
+	if s.Res.TagNamespaceId != nil {
+		s.D.Set("tag_namespace_id", *s.Res.TagNamespaceId)
+	}
+
+	if s.Res.TimeCreated != nil {
+		s.D.Set("time_created", s.Res.TimeCreated.String())
+	}
+
+	if s.Res.Value != nil {
+		s.D.Set("value", *s.Res.Value)
+	}
+
+	return nil
+}